+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:stylable_calendar/stylable_calendar.dart';
 
 void main() => runApp(MyApp());